backend/requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
ccxt==4.3.15
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozendict==2.4.2
frozenlist==1.4.1
gunicorn==22.0.0
html5lib==1.1
idna==3.7
joblib==1.4.2
lxml==5.2.1
multidict==6.0.5
multitasking==0.0.11
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peewee==3.17.3
pycares==4.4.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.28
requests==2.31.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.38
